==> buli_scorers/03_exploringSpielerData.ipynb <==
# Fill in missing values: a case study with DFB data

A data science job may seem a glamorous one for very smart guys with eyeglasses and little bit nerds, but it is not. Among data scientist tasks there are quite a few that are "dirty", for example data cleaning. In particular, they must deal with missing data day in and day out, which becomes a painful but very important duty. Just remember that many machine learning algorithms do not know what to do with missing data.

So, in this notebook we are going to carry out some data cleaning tasks in a small dataset that was previously retrieved from the Deutscher Fußball Bund's archives. This dataset contains some personal information about players that manage to score at least a goal in 1.Bundesliga.

Firstly, let us import the `pandas` library and load the data. In order to reproduce the next steps locally you may download the CSV file here, and remember to change `SCORERS_PATH` to your local file.

# imports
import pandas as pd

# load data
SCORERS_PATH = '/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/spielers.csv'
df_spielers = pd.read_csv(SCORERS_PATH)

Once we have loaded the data, it is recommendable to make some explorations to gain insight about the data and the undertaking we got ahead.

df_spielers.info()

Notice there are some missing values in columns `Nation` and `Nationalelf`, 14 to be exact. In addition, there is no need of the column `DFB_id` and `link` since they are OK by construction then, we can drop them to focus in the other features. Before we dive into the missing values question, let us take a look to our dataframe.

df_spielers.drop(columns=['DFB_id', 'link'], inplace=True)
df_spielers.head(15)

After a visual inspection we infer that `Nation` has lists of strings as values, whilst `Nationalelf` has strings. Recall that `Nation` stores which countries a footballer may represent and `Nationalelf` the chosen team. 

Regarding null values, we have two choices to fill the missing data: we can try to figure out an automatized method, perhaps relying on Transfermarkt, or we can do it by hand. Taking into account that we only have 14 "problematic" rows, it is feasible to complete the information by hand.

Let us take a closer look to the rows we are interested in.

mask_null_values = df_spielers['Nation'].isnull()
df_spielers[mask_null_values]

Let us follow a naive approach: we will construct a couple of lists with the missing data first, and then fill the blanks. The information regarding FIFA eligibility and national teams was retrieved from the specialized portal Transfermarkt.

nation_lst = [
    ['Deutschland', 'Kroatien'], ['Portugal', 'Guinea-Bissau'], ['Japan'], ['Ghana'], 
    ['Griechenland', 'Deutschland'], ['Norway'], ['Argentinien', 'Italien'], ['Norway'], ['Italien'], ['Senegal', 'Deutschland'], ['Tunesien'], ['Japan'], ['Schweiz', 'Kosovo'], ['Ungarn']
]
df_spielers[mask_null_values]['Nation'] = nation_lst

What happened here? `Pandas` has issued a warning message because we have intended to do something that it may result in unwanted behavior. The essence is that the result of the command `df_spielers[mask_null_values]['Nation']` is not a "hard copy" of the dataframe's section but a "shallow copy". In other words, we are not watching the real data but an image. A deeper explanation for those interested may be find [here](https://realpython.com/pandas-settingwithcopywarning/).

Then, to avoid this warning we are going to use the `.loc` method to access the real data and fill in the missing values, but first let us construct a list for `Nationalelf` and after we will use a loop to complete the task.

nationalelf_lst = [
    'Deutschland', 'Portugal', 'Japan', 'Ghana', 'Griechenland', 'Norway', 'Argentina', 'Norway', 'Italien', 'Senegal',
     'Tunesien', 'Japan', 'Schweiz', 'Ungarn'
]

counter = 0
for k in range(len(mask_null_values)):
    if mask_null_values[k]:
        df_spielers.loc[k, 'Nation'] = nation_lst[counter]
        df_spielers.loc[k, 'Nationalelf'] = nationalelf_lst[counter]
        counter+=1

Let us check if the job was done correctly:

df_spielers[mask_null_values]

But, there is something interesting in how this data looks like and how the first 15 rows looks like. Notice how in the latter there are single quotation marks in the column `Nation`. Are the same data types stored in the aforementioned column?

type(df_spielers['Nation'][0])

Thus, `pandas` loaded the values in the column `Nation` as strings and not as lists. Consequently, we would need some transformation in order to bring the lists back. The fastest way to accomplish this is to use the function `literal_eval` from the module `ast` which is more efficient than rely on regular expressions.

from ast import literal_eval

df_spielers['Nation'] = df_spielers['Nation'].apply(
    lambda x: literal_eval(x) if isinstance(x, str) else x
)

The condition `isinstance(x, str)` was added to avoid problems in execution time due to the fact that some rows already have lists as values.

df_spielers

Since I consider myself a perfectionist (when I can allow such luxury, of course) there is still one last step: to check if there exist another kind of errors in the data. An in depth revision threw that there were four players with wrong information, they are Raúl Bobadilla, Sébastien Haller, Pál Dardai and Roman Neustädter.

mask_wrongs = [False]*len(df_spielers)
wrong_pos = [526, 632, 871, 1602]
for pos in wrong_pos:
    mask_wrongs[pos] = True

df_spielers[mask_wrongs]

df_spielers.iloc[526, 2] = 'Paraguay'
df_spielers.iloc[632, 2] = "Elfenbeinküste (Côte d'Ivoire)"
df_spielers.iloc[871, 2] = 'Ungarn'
df_spielers.iloc[871, 1] = ['Ungarn']
df_spielers.iloc[1602, 2] = 'Russland'

df_spielers[mask_wrongs]

Fixing the above errors, we could notice that Côte d'Ivoire appears with its German toponym also. Though the official name of this African country is Côte d'Ivoire we will keep its German version for simplicity.

df_spielers['Nationalelf'] = df_spielers['Nationalelf'].apply(
    lambda x: x.split()[0] if 'Elfen' in x else x 
)

mask_jugoslawien = df_spielers['Nationalelf'] == 'Jugoslawien'
df_spielers[mask_jugoslawien]

ex_jugoslawien = ['Bosnien-Herzegowina', 'Serbien', 'Serbien', 'Kroatien', 'Serbien', 'Serbien', 
                  'Kroatien', 'Serbien', 'Serbien', 'Kroatien', 'Kroatien']
ex_jugoslawien_pos = [k for k in range(len(mask_jugoslawien)) if mask_jugoslawien[k]]

counter=0
for pos in ex_jugoslawien_pos:
    df_spielers.iloc[pos, 1] = [ex_jugoslawien[counter]]
    df_spielers.iloc[pos, 2] = ex_jugoslawien[counter]
    counter+=1

df_spielers[mask_jugoslawien]

These mistakes on footballes from the ex-Yugoslavia tell us something more interesting about the East-West relations during the Cold War. If we look at the countries in the list there are no players from the USSR, while there are players from the socialist Yugoslavia.

df_spielers.to_csv('/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/spielers_fixed.csv')
