==> buli_scorers/04_scrapFIFAcode.ipynb <==
# Scrapping FIFA country codes from Wikipedia

One of the goals of the project is to create dashboard to summarize the information retrieved and curated from DFB archives. Then, we will download the FIFA country codes from Wikipedia.

# imports
import requests
import pandas as pd
from bs4 import BeautifulSoup

# wikipedia webpage
url_FIFAcodes = 'https://de.wikipedia.org/wiki/Liste_der_FIFA-Mitglieder'

# define dataframe
df = pd.DataFrame(columns=['Beitritt', 'Staat', 'FIFA code', 'Verband', 'Gründung', 'Konföderation'])

# Create a handle, page, to handle the contents of the website
data = requests.get(url_FIFAcodes).text

# Creating BeautifulSoup object
soup = BeautifulSoup(data, 'html.parser')

# Creating list with the tables
table = soup.find('table')

# Collecting Data
for row in table.tbody.find_all('tr'):
    # Find all data for each column
    columns = row.find_all('td')
        
    if(columns != []):
        beitritt = columns[0].text.strip()
        temp_staat = columns[1].find_all('a')
        if len(temp_staat[-1].text)>2:
            staat = temp_staat[-1].text
        else:
            staat = temp_staat[-2].text
        code = columns[2].text.strip()
        verband = columns[3].text.strip()
        gruendung = columns[4].text.strip()
        konfoederation = columns[5].text.strip()
        
        df.loc[len(df)] = [beitritt, staat, code, verband, gruendung, konfoederation]

df[['Staat', 'FIFA code', 'Konföderation']]

change_county = {
    'Bosnien und Herzegowina': 'Bosnien-Herzegowina',
    'VR China': 'China',
    'Demokratische Republik Kongo': 'DR Kongo',
    'Republik Kongo': 'Kongo'
}

df.to_csv('/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/FIFA_codes.csv', index=False)

# wikipedia webpage
url_FIFA_ISOcodes = 'https://en.wikipedia.org/wiki/Comparison_of_alphabetic_country_codes'

# define dataframe
df_fifa_iso = pd.DataFrame(columns=['Country', 'IOC code', 'FIFA code', 'ISO code'])

# Create a handle, page, to handle the contents of the website
data_countries = requests.get(url_FIFA_ISOcodes).text

# Creating BeautifulSoup object
soup_countries = BeautifulSoup(data_countries, 'html.parser')

# Creating list with the tables
table_countries = soup_countries.find('table')

df_fifa_iso = pd.DataFrame(columns=['Country', 'IOC code', 'FIFA code', 'ISO code'])
# Collecting Data
for row in table_countries.tbody.find_all('tr'):
    # Find all data for each column
    columns = row.find_all('td')
        
    if(columns != []):
        country = columns[1].find('a')['title']
        if len(columns[2].find_all('a'))==0:  ioc_code = columns[2].text.strip()
        else: ioc_code = ''
        if len(columns[3].find_all('a'))==0: fifa_code = columns[3].text.strip()
        else: fifa_code =  ''
        if len(columns[4].find_all('a'))==0: iso_code = columns[4].text.strip()
        else: iso_code = ''
        
        df_fifa_iso.loc[len(df_fifa_iso)] = [country, ioc_code, fifa_code, iso_code]

df_fifa_iso.loc[61:90]

df_FIFA = df_fifa_iso[~(df_fifa_iso['FIFA code']=='')]
df_FIFA

df_FIFA.to_csv('/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/FIFA_ISO_codes.csv', index=False)
