==> buli_scorers/05_mergeFIFAcodes.ipynb <==
# Getting country names in German and English

# imports
import pandas as pd

# path to data
FIFA_ISO = '/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/FIFA_ISO_codes.csv'
FIFA_GERMAN = '/mnt/287A29DF7A29AA90/PythonProjects/projects_repo/BuLi_scorers/data/FIFA_codes.csv'

# load data
df_fifa_iso = pd.read_csv(FIFA_ISO)
df_fifa_ger = pd.read_csv(FIFA_GERMAN)

# selecting columns
df_fifa_ger = df_fifa_ger[['Staat', 'FIFA code', 'Konföderation']].copy()
df_fifa_ger

df_fifa = df_fifa_ger.merge(df_fifa_iso, on='FIFA code')
df_fifa
